// 3. FrameWorks/2. Angular/curso-angular/src/app/models/product.model.ts
export interface Product {
    id: string;
    title: string;
    price: number;
    images: string[];
    description: string;
    category: Category;
  }
  export interface UpdateProductDTO extends Partial<CreateProductDTO> {}

  export interface Category {
    id: string;
    name: string;
  }
  export interface CreateProductDTO extends Omit<Product, 'id' | 'category'> {
    categoryId: number;
  }